Trim route terminus links in API response.
Trimming/cropping is done in order to get more accurate geometries and lengths for matched results compared to input route geometries. More accurate results aid in recognising the weakest matches in order to find issues with map-matching algorithm and/or Digiroad modeling compared to Jore3 infrastructure network.

The change does not affect e.g. Jore4 UI since individual links in the API response remain unchanged. In other words, the IDs, traversing directions and geometries of individual links are not affected. Only the compound route geometry and the length calculated based on it are changed.