Fix a bug in map-matching API while snapping stop points to infrastructure links. The root cause was the inaccuracies in Digiroad data.
It is suspected that there are inaccuracies in Digiroad data in such a way that in some cases a distance of a public transport stop from the start of the infrastructure link along which the stop is located is greater than the length of the link. This caused errors in resolving fractional location of stop on link. In these cases the solution is to calculate fractional location manually using JTS which seems to fix most occurrences of the issue.

However, a small number of cases were not fixed. It remains to be seen after rigorous testing whether additional measures need to be taken.