Fix writing failed segments.
Let's make a fixed version of the `GeoPackage.add()` method, which originally
worked, but at some point broke when the GeoTools libraries were updated.

The current version of the method throws a `NullPointException` exception for
a missing `fid` property. Debugging the `gt-geopkg` library has revealed that it
probably has a bug. The fact that the documentation is notoriously incomplete
has not helped in solving the problem.